Remove gpu_arch_match from hipblaslt.h
