+++ library/src/include/auxiliary.hpp
@@ -29,8 +29,26 @@
 #include <hipblaslt/hipblaslt.h>
 #include <iostream>
 #include <regex>
+#include <string_view>
 
-inline bool IsOCPSupported()
+/*! \brief device matches pattern */
+inline
+bool gpu_arch_match(std::string_view gpu_arch, std::string_view pattern)
+{
+    if(!pattern.length())
+    {
+        return true;
+    }
+
+    constexpr char    prefix[]   = "gfx";
+    const std::size_t prefix_len = std::string_view(prefix).length();
+    gpu_arch.remove_prefix(prefix_len);
+    std::regex arch_regex(pattern.data());
+    return std::regex_search(gpu_arch.data(), arch_regex);
+}
+
+inline
+bool IsOCPSupported()
 {
     int             deviceId;
     hipDeviceProp_t deviceProperties;
